# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Food Services Auth",
    "description": "Authentication microservice."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": ""
    }
  ],
  "paths": {
    "/users/": {
      "post": {
        "description": "Endpoint to create User",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        }
      },
      "get": {
        "description": "Endpoint to list Users",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Endpoint to retrive User",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Endpoint to update Users",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Endpoint to delete Users",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "description": "Endpoint to login User",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      }
    },
    "/users/validate-token": {
      "post": {
        "description": "Endpoint to validate User auth token",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Token"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreate": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jhon"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "age": {
            "type": "number",
            "example": 42
          },
          "userName": {
            "type": "string",
            "example": "JohnDoe123"
          },
          "password": {
            "type": "string",
            "example": "@Pas$wrd123!"
          }
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jhon"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "age": {
            "type": "number",
            "example": 42
          },
          "userName": {
            "type": "string",
            "example": "JohnDoe123"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "age": {
            "type": "number",
            "example": 42
          },
          "userName": {
            "type": "string",
            "example": "JohnDoe123"
          }
        }
      },
      "Users": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/User"
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "example": "JohnDoe123"
          },
          "password": {
            "type": "string",
            "example": "@Pas$wrd123!"
          }
        }
      },
      "Token": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIkpXVCJ9.eyJkYXRhIjp7ImlkIjoyLCJmaXJzdE5hbWUiOiJmaXJzdCBuYW1lIiwibGFzdE5hbWUiOiJsYXN0IG5hbWUiLzZXJuYW1lMSIsInBhc3N3b3JkIjoiJDJiJDE1JGVHZ3B0ZS5ycFBtazZPRlEvY3lDRWVNWkowOE9DQVY4VXh5SE9vdlAvaWR5aVRpTzN5YnB1In0sImDY5MCwiZXhwIjoxNzAzNzk0MjkwfQ.Zgd-nGQKBuuia8miCrX0UKFvWQPsz-CPqE"
          }
        }
      }
    }
  }
}